feat: add fitur 1-2

+++ apps/api/src/controllers/address.controllers.ts
@@ -2,54 +2,50 @@ import { Request, Response } from 'express';
 import prisma from '@/prisma';
 import { Address } from '@prisma/client';
 import axios from 'axios';
+const geocodeApiUrl = `https://api.opencagedata.com/geocode/v1/json`;
 
 export class AddressController {
-  async getAllAddresses(req: Request, res: Response): Promise<Response> { ////<--------------------------------
+  async getAllAddresses(req: Request, res: Response): Promise<Response> { 
     try {
       const addresses: Address[] = await prisma.address.findMany();
-
       if (!addresses.length) {
         return res.status(404).send({ error: 'No addresses found' });
       }
-
       return res.status(200).send(addresses);
     } catch (error) {
       return res.status(500).send({ error: 'Error fetching addresses' });
     }
   }
 
-  async getAllAddressByRole(req: Request, res: Response) { ////<--------------------------------
+  async getAllAddressByRole(req: Request, res: Response) { 
     const { role } = req.params;
     if (role === 'outlet') {
       try {
-        const addresses: Address[] = await prisma.address.findMany({
+        const page = parseInt(req.query.page as string) || 1;
+        const limit = parseInt(req.query.limit as string) || 2;
+        const startIndex = (page - 1) * limit;
+        const addresses = await prisma.address.findMany({
           where: { outletId: { not: 0 } },
         });
-
         if (!addresses.length) {
-          return res
-            .status(404)
-            .send({ error: 'No addresses found for outlets' });
+          return res.status(404).send({ error: 'No addresses found for outlets' });
         }
-
-        return res.status(200).send(addresses);
+        const paginatedAddress = addresses.slice(startIndex, startIndex + limit);
+        return res.status(200).send(
+          {status:"ok", message: "Get all outlet addresses successfully", data: paginatedAddress, currentPage: page, totalPages: Math.ceil(addresses.length / limit)}
+      );
       } catch (error) {
-        return res
-          .status(500)
-          .send({ error: 'Error fetching outlet addresses' });
+        return res.status(500).send({ error: 'Error fetching outlet addresses' });
       }
     } else if (role === 'user') {
       try {
         const addresses = await prisma.address.findMany({
           where: { userId: { not: 0 } },
-        });
+        })
 
         if (!addresses.length) {
-          return res
-            .status(404)
-            .send({ error: 'No addresses found for users' });
+          return res.status(404).send({ error: 'No addresses found for users' });
         }
-
         return res.status(200).send(addresses);
       } catch (error) {
         return res.status(500).send({ error: 'Error fetching user addresses' });
@@ -59,88 +55,115 @@ export class AddressController {
     }
   }
 
-  async getAddressById(req: Request, res: Response) { ////<--------------------------------
-    const { id } = req.params;
+  async getOutlesAddressByState(req: Request, res: Response) {
+    const { state } = req.params;
+    try {
+      if (!state) throw "City is required";
+      const addresses = await prisma.address.findMany({
+        where:{state, outletId: {not: 0}},
+      })
+      if(!addresses) throw "Address not found";
+      return res.status(200).send({status:"ok", message: "Get Address By State Successfully", data: addresses});
+    } catch (error) {
+      if (error instanceof Error) {res.status(400).send({status: "error",message: error.message,});}
+      res.status(400).send({ status: "error",message: error,});
+    }
+  }
 
+  async getAddressById(req: Request, res: Response) {
+    const { id } = req.params;
     try {
       const address = await prisma.address.findUnique({
         where: { id: Number(id) },
       });
-
       if (!address) {
         return res.status(404).send({ error: 'Address not found' });
       }
-
-      return res.status(200).send(address);
+      return res.status(200).send({ status: 'ok', message: 'Get Address By Id Successfully', data: address });
     } catch (error) {
-      return res.status(500).send({ error: 'Error fetching address' });
+      return res.status(500).send({ status: 'error', message: 'Error fetching address' });
     }
   }
-
   async createAddress(req: Request, res: Response) {
-    const {
-      userId,
-      outletId,
-      address,
-      city,
-      state,
-      postalCode,
-      country,
-    } = req.body;
-  
-    if (!userId && !outletId) {
-      return res.status(400).send({ error: 'Either userId or outletId must be provided' });
-    }
-  
+    const { phone, address, city, state, country, postalCode, outletId, userId } = req.body;
+    
     try {
-      const apiKey = '671dfa638d605573966053jft24917e';
-  
-      const geoResponse = await axios.get(`https://geocode.maps.co/search`, {
-        params: {
-          q: `${address}, ${city}, ${state}, ${postalCode}, ${country}`,
-          api_key: apiKey,
-        },
-      });
-  
-      const data = geoResponse.data[0];
-  
-      if (!data) {
-        return res.status(404).send({ error: 'Address not found' });
+      if (!userId && !outletId) {
+        return res.status(400).send({ status: "error" ,message: 'Either userId or outletId must be provided' });
       }
-  
-      const { lat, lon, display_name } = data;
-  
+      const geocodeResponse = await axios.get(geocodeApiUrl, {
+        params: {
+          q: `${address}, ${city}, ${state}, ${country}, ${postalCode}`,
+          key: process.env.OPENCAGE_API_KEY,
+          language: 'en',
+        }
+      });
+      if (geocodeResponse.data.length == 0) {
+        return res.status(400).json({ error: 'Address not found.' , geocodeResponse: geocodeResponse.data});
+      }
+      const { lat, lng } = geocodeResponse.data.results[0].geometry;
+      const latitude = parseFloat(lat);
+      const longitude = parseFloat(lng);
+      const checkAddress = await prisma.address.findMany({ where: { userId } });
+      let isMain = checkAddress.length === 0; 
+
       const newAddress = await prisma.address.create({
         data: {
-          userId: userId || null, 
+          userId : userId || null,
           outletId: outletId || null,
-          address: display_name,
-          city: city || '',
-          state: state || '',
-          postalCode: postalCode || '',
-          country: country || '',
-          latitude: parseFloat(lat),
-          longitude: parseFloat(lon),
+          address: address,
+          city: city,
+          state: state,
+          isMain: isMain,
+          postalCode: postalCode,
+          country: country,
+          phone: phone,
+          latitude: latitude,   // Menyimpan latitude
+          longitude: longitude,  // Menyimpan longitude
+          createdAt: new Date(),
+          updatedAt: new Date(),
         },
       });
-  
-      return res.status(201).send(newAddress);
+
+      return res.status(201).json({
+        status: 'ok',
+        message: 'Create Address Successfully',
+        data: newAddress,
+      });
     } catch (error) {
-      console.error('Error creating address:', error);
-      return res.status(500).send({ error: 'Error creating address' });
+      return res.status(500).json({ 
+        status: 'error', 
+        message: error 
+      });
     }
   }
-  
 
+  async getAddresByUserId(req: Request, res: Response) {
+    const { id } = req.params;
+    try {
+      if(!id) throw new Error('User id not found');
+      const addresses = await prisma.address.findMany({
+        where: { userId: Number(id) },
+      })
+      if(!addresses){
+        return res.status(404).send({ status: 'error', message: 'The user has not created an address' });
+      }
+      return res.status(200).send({
+        status: 'ok',
+        message: 'Get All Address By User Id Successfully',
+        data: addresses
+      });
+    } catch (error) {
+      if (error instanceof Error) {res.status(400).send({status: "error",message: error.message,});}
+      res.status(400).send({ status: "error",message: error,});
+    }
+  }
   async createAddressReverse(req: Request, res: Response) {
     const { userId, outletId, latitude, longitude } = req.body;
-  
-    
-    if (!userId && !outletId) {
-      return res.status(400).send({ error: 'Either userId or outletId must be provided' });
-    }
-  
     try {
+      if (!userId && !outletId) {
+        return res.status(400).send({ error: 'Either userId or outletId must be provided' });
+      }
       const apiKey = '671dfa638d605573966053jft24917e';
   
       const reverseGeoResponse = await axios.get(
@@ -184,60 +207,80 @@ export class AddressController {
     }
   }
   
-
-  async updateAddress(req: Request, res: Response) { ////<--------------------------------
-    const { id } = req.params;
-    const {
-      userId,
-      outletId,
-      address,
-      city,
-      state,
-      postalCode,
-      country,
-      latitude,
-      longitude,
-    } = req.body;
-
+  async updateAddress(req: Request, res: Response) {
+    const { id, address, city, state, postalCode, country, phone } = req.body;
     try {
+      const geocodeResponse = await axios.get(geocodeApiUrl, {
+        params: {q: `${address}, ${city}, ${state}, ${country}, ${postalCode}`,key: process.env.OPENCAGE_API_KEY,language: 'en'}
+      });
+      if (geocodeResponse.data.length == 0) throw new Error('Address not found.')
+      const { lat, lng } = geocodeResponse.data.results[0].geometry;
+      const latitude = parseFloat(lat);
+      const longitude = parseFloat(lng);
+
       const updatedAddress = await prisma.address.update({
         where: { id: Number(id) },
         data: {
-          userId,
-          outletId,
           address,
           city,
           state,
           postalCode,
           country,
+          phone,
           latitude,
-          longitude,
+          longitude
         },
       });
 
       return res.status(200).send(updatedAddress);
     } catch (error: any) {
-      if (error.code === 'P2025') {
-        return res.status(404).send({ error: 'Address not found' });
+      if (error instanceof Error) {
+        return res.status(400).send({status: "error", message: error.message});
       }
-      return res.status(500).send({ error: 'Error updating address' });
+      return res.status(400).send({status: "error",message: error});
     }
   }
 
-  async deleteAddress(req: Request, res: Response) {  ////<--------------------------------
-    const { id } = req.params;
-
+  async setMainAddress(req: Request, res: Response) {
+    const { userId, addressId } = req.body;
     try {
-      await prisma.address.delete({
+      if(!userId) throw new Error('User id not found');
+      if(!addressId) throw new Error('Address id not found');
+      await prisma.address.updateMany({
+        where: { userId: Number(userId) , isMain: true},
+        data: { isMain: false },
+      }) 
+      await prisma.address.update({
+        where: { id: Number(addressId) },
+        data: { isMain: true },
+      })
+      return res.status(201).json({
+        status: 'ok',
+        message: 'Set Main Address Successfully',
+      });
+    } catch (error) {
+      if (error instanceof Error) {
+        return res.status(400).send({status: "error", message: error.message});
+      }
+      return res.status(400).send({status: "error",message: error});
+    }
+  }
+
+
+  async deleteAddress(req: Request, res: Response) {
+    const { id } = req.params;
+    try {
+      if(!id) throw new Error('Address id not found');
+      const deletedAddress = await prisma.address.delete({
         where: { id: Number(id) },
       });
-
-      return res.status(204).send();
+      if(!deletedAddress) throw new Error('Address not found');
+      return res.status(200).send({status: 'ok', message: 'Delete Address Successfully',data: deletedAddress});
     } catch (error: any) {
-      if (error.code === 'P2025') {
-        return res.status(404).send({ error: 'Address not found' });
+      if (error instanceof Error) {
+        return res.status(400).send({status: "error", message: error.message});
       }
-      return res.status(500).send({ error: 'Error deleting address' });
+      return res.status(400).send({status: "error",message: error});
     }
   }
 }
